Use parameter bounds as linear mapping function for modulation and move to larger signed values to fix casting issues

+++ Core/Inc/WS2812.h
@@ -80,7 +80,7 @@ void WS2812_SetLEDAdditive(uint16_t index, uint8_t red, uint8_t green, uint8_t b
 void WS2812_SetAllLEDs(uint32_t red, uint32_t green, uint32_t blue);
 void WS2812_ClearLEDs(void);
 void WS2812_FadeAll(uint8_t denominator);
-void WS2812_ShiftLEDs(int8_t shiftAmount);
+void WS2812_ShiftLEDs(int16_t shiftAmount);
 
 // Communication with LED strip
 #ifndef USE_NEW_SEND_FUNCTIONS
@@ -97,7 +97,7 @@ void WS2812_MultiCometEffect(void);
 void WS2812_CometEffect(void);
 void WS2812_SimpleMeterEffect(colorRGB color, uint8_t level, bool flip);
 void WS2812_MirroredMeterEffect(colorRGB color, uint8_t level, bool centered);
-void WS2812_FillRainbow(colorHSV startingColor, int8_t deltaHue);
+void WS2812_FillRainbow(colorHSV startingColor, int16_t deltaHue);
 
 // Utility functions
 void WS2812_SetBackgroundColor(uint8_t red, uint8_t green, uint8_t blue);
